Isolated HTTP parsing code into a distinct class
Fixes #2
