+++ Gemfile
@@ -13,4 +13,5 @@ group :test do
   gem "webmock"
   gem "timecop"
   gem "simplecov"
+  gem "mocha"
 end
