+++ test/test_helper.rb
@@ -1,12 +1,12 @@
+# frozen_string_literal: true
+
 require 'simplecov'
 SimpleCov.start
-
-# frozen_string_literal: true
 
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "meteoalarm"
 
-
 require "minitest/autorun"
 require "webmock/minitest"
+require 'mocha/minitest'
 require "timecop"
